+++ packages/ag_flow_cli/bricks/collection_module/__brick__/repos/{{module_file_base}}_repo.dart
@@ -7,4 +7,16 @@ class {{module_class_prefix}}Repo extends AgBaseRepo {
   final {{module_class_prefix}}Service _service;
 
   Future<AgListPage<dynamic, int>> getPage(int pageKey) => _service.getPage(pageKey);
+{{#generate_add}}
+
+  Future<dynamic> add(dynamic item) => _service.add(item);
+{{/generate_add}}
+{{#generate_update}}
+
+  Future<dynamic> update(dynamic id, dynamic item) => _service.update(id, item);
+{{/generate_update}}
+{{#generate_delete}}
+
+  Future<void> delete(dynamic id) => _service.delete(id);
+{{/generate_delete}}
 }