+++ packages/ag_flow_cli/test/goldens/collection_module/lib/product/repos/products_repo.dart
@@ -8,4 +8,10 @@ class ProductsRepo extends AgBaseRepo {
 
   Future<AgListPage<dynamic, int>> getPage(int pageKey) =>
       _service.getPage(pageKey);
+
+  Future<dynamic> add(dynamic item) => _service.add(item);
+
+  Future<dynamic> update(dynamic id, dynamic item) => _service.update(id, item);
+
+  Future<void> delete(dynamic id) => _service.delete(id);
 }