+++ packages/ag_flow_cli/test/goldens/detail_module/lib/product/repos/product_details_repo.dart
@@ -9,4 +9,10 @@ class ProductDetailsRepo extends AgBaseRepo {
 
   Future<dynamic> getByArgument(ProductDetailsPageArgument argument) =>
       _service.getByArgument(argument);
+
+  Future<dynamic> update(ProductDetailsPageArgument argument, dynamic item) =>
+      _service.update(argument, item);
+
+  Future<void> delete(ProductDetailsPageArgument argument) =>
+      _service.delete(argument);
 }